requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Werkzeug==2.2.3
gunicorn==20.1.0
python-dotenv==1.0.0
# Added for PythonAnywhere compatibility